Check size of base64 encoded string not raw string

+++ modules/setting/setting.go
@@ -293,7 +293,7 @@ var (
 		MaxResponseItems       int
 		DefaultPagingNum       int
 		DefaultGitTreesPerPage int
-		DefaultMaxBlobSize     int64
+		DefaultMaxBlobSize     int
 	}{
 		EnableSwagger:          true,
 		MaxResponseItems:       50,
